General settings logic rewrite
