+++ .gitignore
@@ -1,3 +1,3 @@
-/.idea/
 /build/
 /.gradle/
+/.idea/*
