+++ build.gradle
@@ -1,8 +1,8 @@
 plugins {
     id 'org.jetbrains.grammarkit' version '2021.1.3'
-    id 'org.jetbrains.intellij' version '1.3.1'
+    id 'org.jetbrains.intellij' version '1.5.2'
     id 'com.github.ManifestClasspath' version '0.1.0-RELEASE'
-    id 'org.jetbrains.changelog' version '1.2.0'
+    id 'org.jetbrains.changelog' version '1.3.1'
 }
 
 apply plugin: 'org.jetbrains.grammarkit'
@@ -25,8 +25,8 @@ dependencies {
     testCompileOnly 'org.projectlombok:lombok:1.18.24'
     testAnnotationProcessor 'org.projectlombok:lombok:1.18.24'
 
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.0'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.0'
 }
 
 // See https://github.com/JetBrains/gradle-intellij-plugin/
