+++ gradle.properties
@@ -4,7 +4,7 @@ org.gradle.parallel=true
 org.gradle.jvmargs=--add-exports java.base/jdk.internal.vm=ALL-UNNAMED
 
 #platformVersion=IU-LATEST-EAP-SNAPSHOT
-platformVersion=2021.3
+platformVersion=2022.2.1
 platformSinceBuild=211
 
 pluginVersion=1.5.2
