+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'activetabhighlighter'
+rootProject.name = 'active-tab-highlighter'
