+++ src/main/java/com/tobszarny/intellij/plugin/activetabhighlighter/config/model/Constants.java
@@ -1,0 +1,11 @@
+package com.tobszarny.intellij.plugin.activetabhighlighter.config.model;
+
+public final class Constants {
+    public static final String GROUP = "Highlighter";
+    public static final String EXTERNAL_ID = "HIGHLIGHTER_TAB";
+
+    private Constants() {
+    }
+
+
+}
