+++ src/main/java/com/tobszarny/intellij/plugin/activetabhighlighter/config/model/HighlighterSettingsProjectConfig.java
@@ -36,8 +36,6 @@ import java.awt.*;
         })
 public class HighlighterSettingsProjectConfig implements PersistentStateComponent<PersistentConfig> {
 
-    public static final String GROUP = "Highlighter";
-    public static final String EXTERNAL_ID = "HIGHLIGHTER_TAB";
     private static final Logger LOGGER = Logger.getInstance(HighlighterSettingsProjectConfig.class);
     public HighlightedTabTextAttributesDescription attributesDescription;
     PersistentConfig persistentConfig;
@@ -58,8 +56,8 @@ public class HighlighterSettingsProjectConfig implements PersistentStateComponen
         persistentConfig.background.color = "#AD2E9C";
         TextAttributes attributes = new TextAttributes();
         attributes.setBackgroundColor(persistentConfig.background.toColor());
-        TextAttributesKey textAttributesKey = TextAttributesKey.createTextAttributesKey(EXTERNAL_ID);
-        attributesDescription = new HighlightedTabTextAttributesDescription(GROUP, GROUP, attributes, textAttributesKey, EditorColorsManager.getInstance().getGlobalScheme());
+        TextAttributesKey textAttributesKey = TextAttributesKey.createTextAttributesKey(Constants.EXTERNAL_ID);
+        attributesDescription = new HighlightedTabTextAttributesDescription(Constants.GROUP, Constants.GROUP, attributes, textAttributesKey, EditorColorsManager.getInstance().getGlobalScheme());
     }
 
     @Nullable
