+++ src/main/resources/META-INF/plugin.xml
@@ -1,7 +1,7 @@
 <idea-plugin>
     <id>com.tobszarny.activetabhighlighter.plugin.id</id>
     <name>Active Tab Highlighter</name>
-    <version>1.5.1</version>
+    <version>2.0.0</version>
     <vendor email="tomasz.obszarny.ga@gmail.com" url="https://github.com/tobszarny">tobszarny</vendor>
 
     <description><![CDATA[
